UNIT CONVERTER APPLICATION
This application will aim at helping its users by converting the units from one
another. This app will take input from the user and convert it to the desired
unit and give the result. For example, it will be able to convert a value in
centimeters to Meters or grams to kilograms and vice versa.

To develop this application, Android Studio is required. The knowledge of
XML would be required for the user interface and the Java programming for
the application’s conversion functionality.