TASK 3
CALCULATOR

A calculator will be an easy application if you have just learned Android and
coding for Java. This Application will simply take the input values and the
operation to be performed from the users.
After taking the input it’ll return
the results to them on the screen. This is a really easy application and doesn’t
need use of any particular package.